continuing getting spike metrics from cells
seem to have the code working for spike metrics around the first events
in the trial, need to adapt it to accept events that are not the first
ones in the trial. this will include making sure that the baseline
doesn't contain other stimuli.... or is that unreasonable for the simple
spike analysis?